[FIX] fix bugs [ADD] package-debug-win [ADD] more debug info

+++ app/main.dev.js
@@ -65,7 +65,7 @@ app.on('window-all-closed', () => {
 });
 
 app.on('will-quit', () => {
-  if (process.env.NODE_ENV === 'production') {
+  if (process.env.NODE_ENV === 'production' && process.env.DEBUG_PROD !== 'true') {
     const { exec } = require('child_process');
     exec('shutdown -h now');
   }
@@ -122,7 +122,7 @@ app.on('ready', async () => {
     if (!mainWindow) {
       throw new Error('"mainWindow" is not defined');
     }
-    if (process.env.NODE_ENV === 'production') {
+    if (process.env.NODE_ENV === 'production' && process.env.DEBUG_PROD !== 'true') {
       mainWindow.setKiosk(true); // 只有生产环境才全屏
     }
     if (process.env.START_MINIMIZED) {
